# Optimize MaterialGenerator.kt