--- a/src/main/kotlin/io/github/jedlimlx/supplemental_patches/shaders/Util.kt
+++ b/src/main/kotlin/io/github/jedlimlx/supplemental_patches/shaders/Util.kt
@@ -137,7 +137,7 @@ fun _computeAllPivots(lst: List<Int>, depth: Int, variable: String, builder: Str
 fun removeId(id: String, string: String): String {
     val id = id.replace("minecraft:", "")
     val newString = Regex("(minecraft:)?$id( |\\r\\n|\\n|\\r|$)").replace(string, "")
-    return Regex("\\\\( |\\r\\n|\\n|\\r)\\\\( |\\r\\n|\\n|\\r)( |\\r\\n|\\n|\\r)").replace(newString, "")
+    return Regex("\\\\( |\\r\\n|\\n|\\r)\\\\( |\\r\\n|\\n|\\r)( |\\r\\n|\\n|\\r|$)").replace(newString, "")
 }
 
 fun List<String>.conditions() = this.joinToString(" && ") {