Breaks the property files into more manageable chunks so that regex doesn't have to do so much work modifying them

Here's some results from profiling the material generation methods:
| Method | Before | After |
|---|---|---|
| generateTerrainMaterials | 1.298 min | 2.898 s |
| generateEntityMaterials | 10.34 ms | 2.496 ms |
| generateIrisMaterials | 470.0 ms | 9.924 ms |
| generateTranslucentMaterials | 10.74 s | 430.6 ms |
| generateBlockEntityMaterials | 1.135 s | 89.01 ms |

The resulting property files should have 1:1 functionality, however the new ones will be longer due to ignoring (and not condensing) unmodified entries.

This should work fine for the other branches? I can make more pulls for them if needed.